Added the LICENCE file
As a test of the event class and to determine how it acts with a description that is extremely long I must create a very long peice of text... So this is my extremely long text. I wonder if it will truncate it or whether it will display the whole thing... Very interesting indeed! Meow!!